scripts step in

+++ PalLibrary/include/config.h
@@ -22,7 +22,7 @@ typedef	unsigned char		uint8;
 typedef	unsigned short		uint16;
 #endif
 #ifndef	uint32
-typedef	unsigned long		uint32;
+typedef	unsigned int		uint32;
 #endif
 #ifndef	uint64
 typedef	unsigned long long	uint64;
@@ -34,7 +34,7 @@ typedef	signed char			sint8;
 typedef	signed short		sint16;
 #endif
 #ifndef	sint32
-typedef	signed long			sint32;
+typedef	signed int		sint32;
 #endif
 #ifndef	sint64
 typedef	signed long long	sint64;
